// ui/src/generated/ExportCommand.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ExportRequest } from "./ExportRequest";

export type ExportCommand = { "type": "Render", job: string, request: ExportRequest, } | { "type": "Cancel", job: string, } | { "type": "ReadChunk", token: string, offset: number, length: number, } | { "type": "Release", token: string, };